fix render the list of address and separate address from search screen, make Edit as a screen itself
